Split 14 Go files exceeding 500 lines into smaller files
Mechanical moves of functions/methods into new files with no logic
changes. The only refactoring is in center/model_input.go where
inline switch-case bodies were extracted into named helper methods.
